refactor(cli): extract agent command registration
